extracting hosts functionalities from install command to internal/hosts package
