+++ internal/hosts/cfg_others.go
@@ -1,6 +1,6 @@
 // +build !windows
 // +build !darwin
 
-package install
+package hosts
 
 const hostsFile = "/etc/hosts"
