+++ internal/hosts/hosts_others.go
@@ -1,6 +1,6 @@
 // +build !windows
 
-package install
+package hosts
 
 import (
 	"bytes"
