+++ internal/hosts/hosts_windows.go
@@ -1,6 +1,6 @@
 // +build windows
 
-package install
+package hosts
 
 import (
 	"fmt"
